requirements.txt
asgiref==3.7.2
Django==4.2.5
djangorestframework==3.14.0
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
django-rest-swagger==2.2.0
django-debug-toolbar==4.2.0
psycopg2-binary==2.9.7
drf-yasg==1.21.7
gunicorn==21.2.0
redis==5.0.0
